+++ src/fairness/account/account.hpp
@@ -24,17 +24,17 @@ namespace accounting {
 class account_t {
 public:
     account_t (const std::string &name,
-               bool is_user, uint64_t shares, uint64_t usage);
+               bool is_user, uint64_t shares, double usage);
 
     void set_name (const std::string &name);
     void set_shares (uint64_t shares);
-    void set_usage (uint64_t usage);
+    void set_usage (double usage);
     void set_fshare (double fshare);
 
     const std::string &get_name () const;
     bool is_user () const;
     uint64_t get_shares () const;
-    uint64_t get_usage () const;
+    double get_usage () const;
     double get_fshare () const;
 
     int dprint (std::ostringstream &out) const;
@@ -43,7 +43,7 @@ private:
     std::string m_name = "";
     bool m_is_user = false;
     uint64_t m_shares = 0;
-    uint64_t m_usage = std::numeric_limits<uint64_t>::max ();
+    double m_usage = std::numeric_limits<double>::max ();
     double m_fshare = 0.0f;
 };
